added homework for sem4
