--- src/main/java/ru/gb/software_architecture/homework/sem4/NotCorrectClientIdException.java
@@ -1,7 +1,0 @@
-package ru.gb.software_architecture.homework.sem4;
-
-public class NotCorrectClientIdException extends MyByTicketException{
-    public NotCorrectClientIdException(String message, int customerId) {
-        super(message, customerId);
-    }
-}
