--- src/main/java/ru/gb/software_architecture/homework/sem4/NullTicketException.java
@@ -1,8 +1,0 @@
-package ru.gb.software_architecture.homework.sem4;
-
-public class NullTicketException extends MyByTicketException {
-
-    public NullTicketException(String message) {
-        super(message);
-    }
-}
