--- src/main/java/ru/gb/software_architecture/homework/sem4/SearchTicketException.java
@@ -1,10 +1,0 @@
-package ru.gb.software_architecture.homework.sem4;
-
-import java.util.Date;
-
-public class SearchTicketException extends MyByTicketException{
-
-    public SearchTicketException(String message, int customerId, Date date) {
-        super(message, customerId, date);
-    }
-}
