+++ src/main/java/ru/gb/software_architecture/homework/sem4/data/BusStation.java
@@ -1,9 +1,9 @@
-package ru.gb.software_architecture.homework.sem4;
+package ru.gb.software_architecture.homework.sem4.data;
 
 /**
  * Автобусная станция
  */
-class BusStation {
+public class BusStation {
 
     private final TicketProvider ticketProvider;
