+++ src/main/java/ru/gb/software_architecture/homework/sem4/data/Ticket.java
@@ -1,11 +1,11 @@
-package ru.gb.software_architecture.homework.sem4;
+package ru.gb.software_architecture.homework.sem4.data;
 
 import java.util.Date;
 
 /**
  * Билет
  */
-class Ticket {
+public class Ticket {
 
     private static int counter;
