+++ src/main/java/ru/gb/software_architecture/homework/sem4/data/TicketProvider.java
@@ -1,4 +1,9 @@
-package ru.gb.software_architecture.homework.sem4;
+package ru.gb.software_architecture.homework.sem4.data;
+
+import ru.gb.software_architecture.homework.sem4.exceptions.impl.NotCorrectClientIdException;
+import ru.gb.software_architecture.homework.sem4.exceptions.impl.NullDateException;
+import ru.gb.software_architecture.homework.sem4.exceptions.impl.NullUserCollectionException;
+import ru.gb.software_architecture.homework.sem4.exceptions.impl.SearchTicketException;
 
 import java.util.Collection;
 import java.util.Date;
