+++ src/main/java/ru/gb/software_architecture/homework/sem4/exceptions/impl/NullDateException.java
@@ -1,4 +1,6 @@
-package ru.gb.software_architecture.homework.sem4;
+package ru.gb.software_architecture.homework.sem4.exceptions.impl;
+
+import ru.gb.software_architecture.homework.sem4.exceptions.MyByTicketException;
 
 import java.util.Date;
